v0.8 Redesign the new 1.5 level format

+++ src/utils/json_utils.cpp
@@ -68,14 +68,14 @@ void jsonReadEnumU8(const nlohmann::json& j, const std::string& name, u8& target
 
 void WriteCBOR(std::ostream& out, const nlohmann::json& j){
 	std::vector<std::uint8_t> v_cbor = nlohmann::json::to_cbor(j);
-	u64 size = v_cbor.size();
-	out.write((char*)&size, sizeof(u64));
+	u32 size = (u32)v_cbor.size();
+	out.write((char*)&size, sizeof(u32));
 	out.write((char*)v_cbor.data(), size);
 }
 
 nlohmann::json ReadCBOR(std::istream& in){
-	u64 size=0;
-	in.read((char*)&size, sizeof(u64));
+	u32 size=0;
+	in.read((char*)&size, sizeof(u32));
 
 	std::vector<std::uint8_t> v_cbor;
 	v_cbor.resize(size);
